LearningProject/GRPC.Server/requirements.txt
grpcio==1.28.1
grpcio-tools==1.28.1
protobuf==3.11.3
six==1.14.0
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pytesseract==0.3.2
